updated deps, rearranged file/folder structure, add node-env.d.ts types;
