--- .prettierrc.js
@@ -1,7 +1,0 @@
-module.exports = {
-  bracketSpacing: true,
-  jsxBracketSameLine: true,
-  singleQuote: true,
-  trailingComma: 'all',
-  semi: true
-};
