+++ node-env.d.ts
@@ -1,0 +1,8 @@
+declare namespace NodeJS {
+
+  export interface ProcessEnv {
+    NODE_ENV: 'development' | 'production' | 'test';
+    NODE_PATH?: string;
+
+  }
+}
