+++ package.json
@@ -3,37 +3,37 @@
   "version": "0.0.1",
   "private": true,
   "scripts": {
-    "android": "react-native run-android",
+    "droid": "react-native run-android",
     "ios": "react-native run-ios",
     "start": "react-native start",
-    "test": "jest",
+    "test": "jest --watch",
     "lint": "eslint . --ext .js,.jsx,.ts,.tsx"
   },
   "dependencies": {
-    "@types/react": "^16.9.56",
+    "@types/react": "16.9.56",
     "react": "16.13.1",
-    "react-hook-form": "^6.12.2",
-    "react-native": "0.63.3",
-    "react-native-keyboard-aware-scroll-view": "^0.9.3",
-    "react-router-native": "^5.2.0"
+    "react-hook-form": "6.12.2",
+    "react-native": "^0.72.4",
+    "react-native-keyboard-aware-scroll-view": "0.9.3",
+    "react-router-native": "5.2.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.8.4",
-    "@babel/runtime": "^7.8.4",
-    "@react-native-community/eslint-config": "^1.1.0",
-    "@types/dotenv": "^8.2.0",
-    "@types/jest": "^25.2.3",
-    "@types/react-native": "^0.63.2",
-    "@types/react-test-renderer": "^16.9.2",
-    "@typescript-eslint/eslint-plugin": "^2.27.0",
-    "@typescript-eslint/parser": "^2.27.0",
-    "babel-jest": "^25.1.0",
-    "eslint": "^6.5.1",
-    "jest": "^25.1.0",
-    "metro-react-native-babel-preset": "^0.59.0",
+    "@babel/core": "7.8.4",
+    "@babel/runtime": "7.8.4",
+    "@react-native-community/eslint-config": "1.1.0",
+    "@types/dotenv": "8.2.0",
+    "@types/jest": "25.2.3",
+    "@types/react-native": "0.63.2",
+    "@types/react-test-renderer": "16.9.2",
+    "@typescript-eslint/eslint-plugin": "2.27.0",
+    "@typescript-eslint/parser": "2.27.0",
+    "babel-jest": "25.1.0",
+    "eslint": "6.5.1",
+    "jest": "^29.6.4",
+    "metro-react-native-babel-preset": "0.59.0",
     "react-test-renderer": "16.13.1",
-    "ts-node": "^9.1.1",
-    "typescript": "^3.8.3"
+    "ts-node": "9.1.1",
+    "typescript": "3.8.3"
   },
   "jest": {
     "preset": "react-native",
