+++ src/components/Input.tsx
@@ -4,10 +4,11 @@ import {
   View,
   Text,
   TextInput,
-  StyleSheet,
   TextStyle,
   TextInputProps,
 } from 'react-native';
+
+import { inputStyles } from '../styles/input.style';
 
 interface InputProps extends TextInputProps {
   name: string;
@@ -16,50 +17,23 @@ interface InputProps extends TextInputProps {
   error?: FieldError | undefined;
 }
 
-const Input = forwardRef<any, InputProps>(
-  (props, ref): ReactElement => {
+const Input = forwardRef<any, InputProps>((props, ref): ReactElement => {
     const { label, labelStyle, error, ...inputProps } = props;
 
     return (
-      <View style={styles.container}>
-        {label && <Text style={[styles.label, labelStyle]}>{label}</Text>}
+      <View style={inputStyles.container}>
+        {label && <Text style={[inputStyles.label, labelStyle]}>{label}</Text>}
         <TextInput
           autoCapitalize="none"
           ref={ref}
-          // eslint-disable-next-line react-native/no-inline-styles
-          style={[styles.input, { borderColor: error ? '#fc6d47' : '#c0cbd3' }]}
+          // eslint-disable-next-line react-native/no-inline-inputStyles
+          style={[inputStyles.input, { borderColor: error ? '#fc6d47' : '#c0cbd3' }]}
           {...inputProps}
         />
-        <Text style={styles.textError}>{error && error.message}</Text>
+        <Text style={inputStyles.textError}>{error && error.message}</Text>
       </View>
     );
   },
 );
-
-const styles = StyleSheet.create({
-  container: {
-    marginVertical: 8,
-  },
-  input: {
-    borderStyle: 'solid',
-    borderWidth: 1,
-    borderRadius: 5,
-    paddingVertical: 5,
-    paddingLeft: 5,
-    fontSize: 16,
-    height: 40,
-    color: '#c0cbd3',
-  },
-  label: {
-    paddingVertical: 5,
-    fontSize: 16,
-    fontWeight: 'bold',
-    color: '#c0cbd3',
-  },
-  textError: {
-    color: '#fc6d47',
-    fontSize: 14,
-  },
-});
 
 export default Input;
