+++ src/styles/register.style.ts
@@ -1,0 +1,18 @@
+import { StyleSheet } from "react-native";
+
+export const registerStyles = StyleSheet.create({
+  container: {
+    display: 'flex',
+    justifyContent: 'flex-start',
+  },
+
+  userName: {
+    fontSize: 16,
+  },
+
+  inputs: {
+    height: 40,
+    borderColor: 'gray',
+    borderWidth: 1,
+  },
+});
